Moved HxError && implemented Error trait
